Made updates related to tour nil class error on the Stops index view
In the stops controller, the index method was not previously receiving the set_tour before action. Once index was added to the before_action for set_tour, there was no longer an error on the index view. The problem was that in the absence of a Stop object, @stops.tour is nil.

The syntax to call a Tour name was updated in the Stops/index view