Photos can be uploaded, put more tweaking needed
At present, the application view in layouts has code for 2 divs, with class - 'container-fluid' and class - 'container'. The former works for the photo upload form and I've commented out the other, but this could cause issues for other view that need to be assessed
The photo displays way too big, even though I added post-processing code
The photo upload route stands alone at present, but needs to be integrated with the other create views and routes that exist
Importantly, I hard-coded id's for org, tour, and stop, for test purposes. This is only a temporary fix